+++ odoo/local-src/sf_sale_invoicing/models/__init__.py
@@ -2,3 +2,4 @@
 # Part of sensefly.
 from . import sale
 from . import invoice
+from . import account_payment_term